+++ colormath/src/commonMain/kotlin/com/github/ajalt/colormath/internal/Constants.kt
@@ -1,5 +1,7 @@
 package com.github.ajalt.colormath.internal
 
+import kotlin.native.concurrent.SharedImmutable
+
 // Constants used in LAB and LUV conversions.
 // http://www.brucelindbloom.com/index.html?LContinuity.html
 /** ϵ = (6/29)^3 */
@@ -13,10 +15,12 @@ internal const val CIE_E_times_K = 8.0
 
 /** The CIECAM02 transform matrix for XYZ -> LMS */
 // https://en.wikipedia.org/wiki/CIECAM02#CAT02
+@SharedImmutable
 internal val CAT02_XYZ_TO_LMS = Matrix(
     +0.7328f, +0.4296f, -0.1624f,
     -0.7036f, +1.6975f, +0.0061f,
     +0.0030f, +0.0136f, +0.9834f,
 )
 
+@SharedImmutable
 internal val CAT02_LMS_TO_XYZ = CAT02_XYZ_TO_LMS.inverse()